third_party/blink/web_tests/external/wpt/css/css-link-params/link-parameters-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property link-parameters value 'none'
  assert_true: link-parameters doesn't seem to be supported in the computed style expected true got false
[FAIL] Property link-parameters value 'param(--a, orange)'
  assert_true: link-parameters doesn't seem to be supported in the computed style expected true got false
[FAIL] Property link-parameters value 'param(--a, orange), param(--b, blue)'
  assert_true: link-parameters doesn't seem to be supported in the computed style expected true got false
[FAIL] Property link-parameters value 'param(--a, )'
  assert_true: link-parameters doesn't seem to be supported in the computed style expected true got false
[FAIL] Property link-parameters value 'param(--a, ), param(--b, )'
  assert_true: link-parameters doesn't seem to be supported in the computed style expected true got false
[FAIL] Property link-parameters value 'param(--a)'
  assert_true: link-parameters doesn't seem to be supported in the computed style expected true got false
[FAIL] Property link-parameters value 'param(--a), param(--b)'
  assert_true: link-parameters doesn't seem to be supported in the computed style expected true got false
Harness: the test ran to completion.
